bigdata-platform/code/common/python-utils/requirements.txt
sqlalchemy>=2.0
pymysql>=1.1.0
psycopg2-binary>=2.9.9
kafka-python>=2.0.2
hdfs>=2.7.3
pyarrow>=14.0
pyyaml>=6.0
python-dateutil>=2.8.2
